Split provider and context
