+++ src/front-end/src/index.js
@@ -4,7 +4,7 @@ import './index.css'
 import App from './App'
 import reportWebVitals from './reportWebVitals'
 import 'bootstrap/dist/css/bootstrap.min.css'
-import {TodoProvider} from "./contexts/TodoContext";
+import TodoProvider from './providers/TodoProvider';
 
   const root = createRoot(document.getElementById('root'))
   root.render(
